move docs to view
